+++ framework/graphs/petsc_graph_partitioner.h
@@ -14,9 +14,9 @@ class PETScGraphPartitioner : public GraphPartitioner
 public:
   explicit PETScGraphPartitioner(const InputParameters& params);
 
-  std::vector<int64_t> Partition(const std::vector<std::vector<uint64_t>>& graph,
-                                 const std::vector<Vector3>& centroids,
-                                 int number_of_parts) override;
+  std::vector<int> Partition(const std::vector<std::vector<uint64_t>>& graph,
+                             const std::vector<Vector3>& centroids,
+                             int number_of_parts) override;
 
 protected:
   const std::string type_;
