+++ framework/mesh/mesh_generator/mesh_generator.h
@@ -54,19 +54,18 @@ protected:
    * Builds a cell-graph and executes the partitioner that assigns cell partition ids based on the
    * supplied number of partitions.
    */
-  std::vector<int64_t> PartitionMesh(const UnpartitionedMesh& input_umesh,
-                                     int num_partitions) const;
+  std::vector<int> PartitionMesh(const UnpartitionedMesh& input_umesh, int num_partitions) const;
 
   /// Executes the partitioner and configures the mesh as a real mesh.
   std::shared_ptr<MeshContinuum> SetupMesh(const std::shared_ptr<UnpartitionedMesh>& input_umesh,
-                                           const std::vector<int64_t>& cell_pids) const;
+                                           const std::vector<int>& cell_pids) const;
 
   /// Determines if a cells needs to be included as a ghost or as a local cell.
   bool CellHasLocalScope(int location_id,
                          const UnpartitionedMesh::LightWeightCell& lwcell,
                          uint64_t cell_global_id,
                          const std::vector<std::set<uint64_t>>& vertex_subscriptions,
-                         const std::vector<int64_t>& cell_partition_ids) const;
+                         const std::vector<int>& cell_partition_ids) const;
 
   const double scale_;
   const bool replicated_;
@@ -81,13 +80,13 @@ protected:
   /// Converts a light-weight cell to a real cell.
   static std::unique_ptr<Cell> SetupCell(const UnpartitionedMesh::LightWeightCell& raw_cell,
                                          uint64_t global_id,
-                                         uint64_t partition_id);
+                                         int partition_id);
 
   static void ComputeAndPrintStats(const std::shared_ptr<MeshContinuum>& grid);
 
   /// Broadcasts PIDs to other locations.
   static void
-  BroadcastPIDs(std::vector<int64_t>& cell_pids, int root, const mpi::Communicator& communicator);
+  BroadcastPIDs(std::vector<int>& cell_pids, int root, const mpi::Communicator& communicator);
 
 private:
 private:
@@ -99,7 +98,7 @@ private:
    * finds a partition with zero cells.
    * \todo Explore more robust partitioners that can better distribute cells across available PEs.
    */
-  static void RebalancePartitions(std::vector<int64_t>& cell_pids, int num_partitions);
+  static void RebalancePartitions(std::vector<int>& cell_pids, int num_partitions);
 };
 
 } // namespace opensn
